requirements.txt
faiss-cpu==1.7.3
transformers==4.33.2
torch==2.0.1
